Moved `LevelFinishedEventArgs` to `Utilities` namespace.
