+++ Beat Saber Utils/Beat Saber Utils.csproj
@@ -124,7 +124,7 @@
     <Compile Include="Gameplay\LevelData.cs" />
     <Compile Include="Gameplay\ScoreSubmission.cs" />
     <Compile Include="Gameplay\UserInfo.cs" />
-    <Compile Include="LevelFinishedEventArgs.cs" />
+    <Compile Include="Utilities\LevelFinishedEventArgs.cs" />
     <Compile Include="Plugin.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
