+++ Beat Saber Utils/Gameplay/ScoreSubmission.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using BS_Utils.Utilities;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
